test(canon): relocate optional listener test to stay under length guard
